# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ToDo doc API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/": {},
        "/api/todos": {
            "get": {
                "tags": [
                    "ToDo"
                ],
                "summary": "Список задач",
                "operationId": "7a0197d175f06353dce08a2e64a5756c",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Some title"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Some description"
                                                    },
                                                    "completed": {
                                                        "type": "boolean",
                                                        "example": "false"
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": "1"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-06T21:59:37.000000"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-11-06T21:59:37.000000"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ToDo"
                ],
                "summary": "Создание новой задачи",
                "operationId": "b1e4acf74bee6f5dbaed653625a10490",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Some title"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Some description"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Some title"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Some description"
                                        },
                                        "completed": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-06T21:59:37.000000"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-06T21:59:37.000000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/todos/{todo}": {
            "get": {
                "tags": [
                    "ToDo"
                ],
                "summary": "Единичная задача",
                "operationId": "5b763f64f916dedd84734399f272a8c5",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Some title"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Some description"
                                        },
                                        "completed": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-06T21:59:37.000000"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-06T21:59:37.000000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ToDo"
                ],
                "summary": "Удалить задачу",
                "operationId": "0ea61052c5938a06c69b90f9c2e5aeed",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "ToDo"
                ],
                "summary": "Обновление задачи",
                "operationId": "e6ab44ccea2c1a4cd393d898a8981ea5",
                "parameters": [
                    {
                        "name": "todo",
                        "in": "path",
                        "description": "ID задачи",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Edited title"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Edited description"
                                            },
                                            "completed": {
                                                "type": "boolean",
                                                "example": "true"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Edited title"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Edited description"
                                        },
                                        "completed": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-06T21:59:37.000000"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-11-06T21:59:37.000000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "ToDo",
            "description": "ToDo"
        }
    ]
}